Day 9: When real math doesn't work, one resorts to stupid calculations.

+++ AdventOfCode/Utilities.cs
@@ -24,7 +24,7 @@ public static class Utilities
 
     public static int Product(this IEnumerable<int> source)
     {
-        return source.Aggregate((int)1, (x, y) => x * y);
+        return source.Aggregate(1, (x, y) => x * y);
     }
 
     public static long Product(this IEnumerable<long> source)
